requirements.txt
fastapi
uvicorn[standard]
pydantic
weaviate-client
requests
numpy
sentence-transformers
torch
groq
